extracted other variables to enums and fixed file names to start with lowercase
