+++ motorlib/enums/units/timeUnit.py
@@ -3,8 +3,5 @@ from enum import Enum
 
 # Python 3.11 supports `StrEnum` that would make this a bit more concise to write
 # https://docs.python.org/3/library/enum.html#enum.StrEnum
-class MassUnit(str, Enum):
-    KILOGRAM = 'kg'
-    GRAM = 'g'
-    POUND = 'lb'
-    OUNCE = 'oz'
+class TimeUnit(str, Enum):
+    SECOND = 's'
