+++ motorlib/grains/custom.py
@@ -4,7 +4,7 @@ import skimage.draw as draw
 
 from ..enums.simAlertLevel import SimAlertLevel
 from ..enums.simAlertType import SimAlertType
-from ..enums.units.LengthUnit import LengthUnit
+from ..enums.units.lengthUnit import LengthUnit
 from ..grain import FmmGrain
 from ..properties import PolygonProperty, EnumProperty
 from ..simResult import SimAlert
