+++ motorlib/grains/dGrain.py
@@ -1,7 +1,7 @@
 """D Grain submodule"""
 from ..enums.simAlertLevel import SimAlertLevel
 from ..enums.simAlertType import SimAlertType
-from ..enums.units.LengthUnit import LengthUnit
+from ..enums.units.lengthUnit import LengthUnit
 from ..grain import FmmGrain
 from ..properties import FloatProperty
 from ..simResult import SimAlert
