+++ motorlib/grains/finocyl.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from ..enums.simAlertLevel import SimAlertLevel
 from ..enums.simAlertType import SimAlertType
-from ..enums.units.LengthUnit import LengthUnit
+from ..enums.units.lengthUnit import LengthUnit
 from ..grain import FmmGrain
 from ..properties import FloatProperty, IntProperty
 from ..simResult import SimAlert
