+++ motorlib/motor.py
@@ -3,9 +3,10 @@ from .enums.multiValueChannels import MultiValueChannels
 from .enums.simAlertLevel import SimAlertLevel
 from .enums.simAlertType import SimAlertType
 from .enums.singleValueChannels import SingleValueChannels
-from .enums.units.LengthUnit import LengthUnit
-from .enums.units.MassFluxUnit import MassFluxUnit
-from .enums.units.PressureUnit import PressureUnit
+from .enums.units.lengthUnit import LengthUnit
+from .enums.units.massFluxUnit import MassFluxUnit
+from .enums.units.pressureUnit import PressureUnit
+from .enums.units.timeUnit import TimeUnit
 from .grains import grainTypes
 from .nozzle import Nozzle
 from .propellant import Propellant
@@ -28,7 +29,7 @@ class MotorConfig(PropertyCollection):
         # Simulation
         self.props['burnoutWebThres'] = FloatProperty('Web Burnout Threshold', LengthUnit.METER, 2.54e-5, 3.175e-3)
         self.props['burnoutThrustThres'] = FloatProperty('Thrust Burnout Threshold', '%', 0.01, 10)
-        self.props['timestep'] = FloatProperty('Simulation Timestep', 's', 0.0001, 0.1)
+        self.props['timestep'] = FloatProperty('Simulation Timestep', TimeUnit.SECOND, 0.0001, 0.1)
         self.props['ambPressure'] = FloatProperty('Ambient Pressure', PressureUnit.PASCAL, 0.0001, 102000)
         self.props['mapDim'] = IntProperty('Grain Map Dimension', '', 250, 2000)
         self.props['sepPressureRatio'] = FloatProperty('Separation Pressure Ratio', '', 0.001, 1)
