+++ uilib/enums/fileAction.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+# Python 3.11 supports `StrEnum` that would make this a bit more concise to write
+# https://docs.python.org/3/library/enum.html#enum.StrEnum
+class FileAction(str, Enum):
+    APPEND = 'Append'
+    OVERWRITE = 'Overwrite'
